+++ app/jenkins-for-jira-ui/src/__mocks__/getAllJenkinsServers.ts
@@ -1,0 +1,1 @@
+export const getAllJenkinsServers = jest.fn(() => Promise.resolve([]));
